src: Update JobAssignmentView
