+++ src/frontend/src/components/home/jobs/management/EmployeeJobCard.vue
