+++ src/frontend/src/components/home/jobs/management/ManagerJobCard.vue
